Add files via upload
This model serves as a baseline model for the subclassing CNN.  Transfer learning upload is forthcoming, as are the patches which have been extracted.   50 X 50 patches have been extracted and padded for ingestion by the CNN.  The FSL annotations for MFs and MLFs have been combined, and figures which were skipped have been dropped entirely.   Accuracy is low, as can perhaps be expected without transfer learning, and sits at about 37%.